+++ src/actions/index.js
@@ -1,15 +1,11 @@
 import axios from 'axios';
-
-const API_URL = `http://localhost:3000/api/addons`
+import io from 'socket.io-client';
 
 export const GET_RESULTS = 'GET_RESULTS';
 
-export function getResults() {
-    // Request is currently a promise
-    const request = axios.get(API_URL);
-    // Returns an action
+export function getResults(rows) {
     return {
         type: GET_RESULTS,
-        payload: request
+        payload: rows
     }
 }
